WebSocket/Abhilekh/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
greenlet==3.0.3
